requirements.txt
requests==2.31.0
selenium==4.18.1
python-dotenv==1.0.1
crewai>=0.30.11
crewai-tools>=0.2.6
pydantic>=2.0